refactor: move some plotting utils out
